wbackend/requirements.txt
coreapi==2.2.4
coreschema==0.0.4
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.30.0
django-cors-headers==2.0.2
django-filter==1.0.1
django-rest-auth==0.8.0
djangorestframework==3.5.4
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
oauthlib==2.0.0
openapi-codec==1.1.3
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
simplejson==3.8.2
six==1.10.0
uritemplate==3.0.0
wheel==0.24.0
